Convert links and joints to USD (#25)
Converts the robot structure (links and joints) and physical properties, but not the meshes/materials

Also implemented a planar joint define function. We may want to contibute this back to usd-exchange at some point.

URDF transmission elements are now considered undefined as the XML for them schema is unstable/unsupported.